requirements.txt
paddlepaddle==2.5.0
paddleocr==2.6.1.3
opencv-python==4.8.0.76
pillow==10.0.0
PyQt5==5.15.9
img2pdf==0.4.4
pyinstaller==5.13.0
numpy==1.24.3 